Upgraded to OpenGL 3.2
+ Using GL3W instead of GLEW
+ Support for integer textures added
+ Shaders now default to #version 130
+ PixelFormat class
+ Upgraded all libraries to VS2015

+++ source/common/engine.cpp
@@ -234,7 +234,7 @@ int Engine::run()
 			if(m_drawFunc)
 			{
 				const double alpha = accumulator / dt;
-				m_drawFunc(m_graphics->s_graphicsContext, alpha);
+				m_drawFunc(m_graphics->s_graphicsContext, (float)alpha);
 				m_graphics->swapBuffers();
 			}
